Move Cisco IOS connection under SSH transport
