+++ test/unit/transports/ssh_test.rb
@@ -5,10 +5,10 @@ require 'train/transports/ssh'
 
 describe 'ssh transport' do
   let(:cls) do
-      plat = Train::Platforms.name('mock').in_family('linux')
-      plat.add_platform_methods
-      Train::Platforms::Detect.stubs(:scan).returns(plat)
-      Train::Transports::SSH
+    plat = Train::Platforms.name('mock').in_family('linux')
+    plat.add_platform_methods
+    Train::Platforms::Detect.stubs(:scan).returns(plat)
+    Train::Transports::SSH
   end
   let(:conf) {{
     host: rand.to_s,
